web/requirements.txt
Django==2.0.1
Pillow==5.0.0
PyJWT==1.5.3
certifi==2017.11.5
chardet==3.0.4
django-filter==1.1.0
django-filters==0.2.1
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
idna==2.6
numpy==1.14.0
opencv-python==3.4.0.12
pip==9.0.1
psycopg2==2.7.3.2
pytz==2017.3
requests==2.18.4
setuptools==38.4.0
wheel==0.30.0
gunicorn==19.6.0
uwsgi==2.0.15